--- apps/web/lib/actions/workspaces.tsx
@@ -1,13 +1,0 @@
-"use server";
-
-import callBackend from "@/lib/call-backend";
-
-export async function deleteWorkspace(workspaceId: string) {
-  return callBackend(
-    `/api/workspaces/${workspaceId}`,
-    {
-      method: "DELETE",
-    },
-    "Workspace deleted successfully.",
-  );
-}
